move testing from tb-utils to tests and some reorg
