+++ tests/Cargo.toml
@@ -9,12 +9,14 @@ rust-version.workspace = true
 async-trait = { workspace = true }
 cliquenet = { path = "../cliquenet", features = ["turmoil"] }
 committable = { workspace = true }
+crossbeam-queue = { workspace = true }
 bytes = { workspace = true }
 futures = { workspace = true }
 multisig = { path = "../multisig" }
 portpicker = { workspace = true }
 rand = { workspace = true }
 sailfish = { path = "../sailfish", features = ["test"] }
+sailfish-types = { path = "../sailfish-types" }
 serde = { workspace = true }
 timeboost = { path = "../timeboost" }
 timeboost-utils = { path = "../timeboost-utils", features = ["test"] }
