+++ tests/src/tests.rs
@@ -1,3 +1,4 @@
 pub mod consensus;
 pub mod network;
+pub mod rbc;
 pub mod timeboost;
