+++ tests/src/tests/network/external.rs
@@ -6,12 +6,12 @@ use cliquenet::{Network, NetworkMetrics};
 use multisig::PublicKey;
 use sailfish::Coordinator;
 use sailfish::rbc::{Rbc, RbcConfig};
-use timeboost_utils::test::message_interceptor::NetworkMessageInterceptor;
-use timeboost_utils::test::testnet::{MsgQueues, TestNet};
 use tokio::task::JoinSet;
 
 use crate::Group;
 use crate::prelude::*;
+use crate::tests::network::message_interceptor::NetworkMessageInterceptor;
+use crate::tests::network::testnet::{MsgQueues, TestNet};
 use crate::tests::network::{TaskHandleResult, TestCondition, TestableNetwork};
 
 pub struct BasicNetworkTest {
