+++ tests/src/tests/network/internal/net.rs
@@ -29,6 +29,8 @@ pub struct Conn<T> {
 
 /// A network event.
 #[derive(Clone, Debug)]
+#[allow(unused)]
+
 pub enum Event<T> {
     Unicast {
         src: PublicKey,
