--- timeboost-utils/src/test.rs
@@ -1,3 +1,0 @@
-pub mod message_interceptor;
-pub mod net;
-pub mod testnet;
